Update cpptoml dependency.
We can now use an include_subdirectory() and link_library() to inherit the
compiler checks (and include paths) in the cpptoml repository without
having to re-write them ourselves.

This change requires CMake 3.0.0 or greater.

+++ src/lm/language_model.cpp
@@ -29,7 +29,7 @@ language_model::language_model(const std::string& config_file)
 {
     auto config = cpptoml::parse_file(config_file);
     auto group = config.get_group("language-model");
-    auto nval = group->get_as<size_t>("n-value");
+    auto nval = group->get_as<int64_t>("n-value");
     if(!nval)
         throw std::runtime_error{"no n-value specified in language-model group"};
